requirements.txt
aiogram
Pillow
ImageHash
shlex
aiohttp